reflect: move folder structure around to be a little better
