+++ librashader-reflect/Cargo.toml
@@ -25,7 +25,7 @@ librashader-presets = { path = "../librashader-presets", version = "0.2.0-beta.9
 
 spirv_cross = { package = "librashader-spirv-cross", version = "0.23", optional = true }
 
-naga = { version = "0.19.0", features = ["spv-in", "wgsl-out"], optional = true }
+naga = { version = "0.19.0", features = ["spv-in", "spv-out"], optional = true }
 rspirv = { version = "0.12.0", optional = true }
 spirv = { version = "0.3.0", optional = true}
 
@@ -39,9 +39,10 @@ version = "0.4"
 optional = true
 
 [features]
-default = ["cross", "wgsl", "serialize"]
-dxil = ["cross", "spirv-to-dxil"]
-wgsl = ["cross", "naga", "spirv", "rspirv"]
-cross = [ "spirv_cross", "spirv_cross/glsl", "spirv_cross/hlsl" ]
+default = ["cross", "naga", "serialize"]
+dxil = ["spirv_cross/hlsl", "spirv-to-dxil"]
+wgsl = ["cross", "naga/wgsl-out", "spirv", "rspirv"]
+cross = [ "spirv_cross", "spirv_cross/glsl", "spirv_cross/hlsl", "spirv_cross/msl" ]
+naga = [ "wgsl" ]
 serialize = [ "serde" ]
-msl = [ "cross", "naga" ]
+msl = [ "spirv_cross/msl", "naga/msl-out" ]
