+++ librashader-reflect/src/back/glsl.rs
@@ -1,13 +1,14 @@
-use crate::back::targets::{GLSL, HLSL};
+use crate::back::targets::GLSL;
 use crate::back::{CompileShader, CompilerBackend, FromCompilation};
 use crate::error::ShaderReflectError;
 use crate::front::SpirvCompilation;
-use crate::reflect::cross::{CompiledProgram, GlslReflect, HlslReflect, SpirvCross};
+use crate::reflect::cross::{CompiledProgram, SpirvCross};
 use crate::reflect::ReflectShader;
 
 /// The GLSL version to target.
 pub use spirv_cross::glsl::Version as GlslVersion;
 
+use crate::reflect::cross::glsl::GlslReflect;
 /// The HLSL shader model version to target.
 pub use spirv_cross::hlsl::ShaderModel as HlslShaderModel;
 
@@ -31,28 +32,6 @@ impl FromCompilation<SpirvCompilation, SpirvCross> for GLSL {
     ) -> Result<CompilerBackend<Self::Output>, ShaderReflectError> {
         Ok(CompilerBackend {
             backend: GlslReflect::try_from(&compile)?,
-        })
-    }
-}
-
-/// The context for a HLSL compilation via spirv-cross.
-pub struct CrossHlslContext {
-    /// The compiled HLSL program.
-    pub artifact: CompiledProgram<spirv_cross::hlsl::Target>,
-}
-
-impl FromCompilation<SpirvCompilation, SpirvCross> for HLSL {
-    type Target = HLSL;
-    type Options = Option<HlslShaderModel>;
-    type Context = CrossHlslContext;
-    type Output = impl CompileShader<Self::Target, Options = Self::Options, Context = Self::Context>
-        + ReflectShader;
-
-    fn from_compilation(
-        compile: SpirvCompilation,
-    ) -> Result<CompilerBackend<Self::Output>, ShaderReflectError> {
-        Ok(CompilerBackend {
-            backend: HlslReflect::try_from(&compile)?,
         })
     }
 }
