+++ librashader-runtime-gl/src/gl/gl3/compile_program.rs
@@ -3,7 +3,7 @@ use crate::error::FilterChainError;
 use crate::gl::CompileProgram;
 use crate::util;
 use gl::types::{GLint, GLuint};
-use librashader_reflect::back::cross::CrossGlslContext;
+use librashader_reflect::back::glsl::CrossGlslContext;
 use librashader_reflect::back::ShaderCompilerOutput;
 use spirv_cross::spirv::Decoration;
